packages/ursa-major/ursa-major-0.4.2.tar.gz/ursa-major-0.4.2/ursa_major/handlers/remove_module.py
# -*- coding: utf-8 -*-
# Copyright (c) 2018  Red Hat, Inc.
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in all
# copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.
#
# Written by Chenxiong Qi <cqi@redhat.com>
#            Qixiang Wan <qwan@redhat.com>

from ursa_major.handlers.base import BaseHandler
from ursa_major.logger import log


class RemoveModuleHandler(BaseHandler):
    def run(self):
        self.dry_run = self.args.dry_run

        self.tag = self.args.tag
        self.module_name = self.args.module_name
        self.module_stream = self.args.module_stream
        self.module_requires = {}
        if self.args.module_requires is not None:
            self.module_requires = dict(self.args.module_requires)
        if self.args.module_buildrequires is not None:
            self.module_buildrequires = dict(self.args.module_buildrequires)
        else:
            self.module_buildrequires = {}
        self.tag_config_file = self.args.tag_config_file

        self.connect_koji()
        self.connect_mbs()

        self.load_tag_config()
        if self.tag not in self.tag_config.keys():
            log.warning("Tag %s is not in tag config file %s, nothing to change",
                        self.tag, self.tag_config_file)
            return

        try:
            tag_inheritance = self.koji.get_inheritance_data(self.tag)
        except Exception as e:
            raise RuntimeError("Unable to get inheritance data of tag {}:\n"
                               "{!s}".format(self.tag, e))

        inheritance_tags = [t['name'] for t in tag_inheritance]

        log.info("Querying modules of '%s:%s' from MBS with requires: %r and "
                 "buildrequires: %r",
                 self.module_name, self.module_stream, self.module_requires,
                 self.module_buildrequires)

        modules = self.mbs.get_modules(
            name=self.module_name,
            stream=self.module_stream,
            requires=self.module_requires,
            buildrequires=self.module_buildrequires)

        module_tags = [m['koji_tag'] for m in modules]
        match_tags = list(set(inheritance_tags) & set(module_tags))
        if match_tags:
            log.info("Found match module tags under %s:\n%s", self.tag, match_tags)
            log.info("You may want to remove these module tags from inheritance by manual later")
        else:
            log.info("No match tag found in tag %s's inheritance data", self.tag)

        module_configs = self.tag_config[self.tag].get('modules', [])

        found_config = None
        for config in module_configs:
            if all((
                self.module_name == config['name'],
                self.module_stream == config['stream'],

                # NOTE: we only find config that matches module_requires exactly
                self.module_requires == config.get('requires', {}),
                # Same as requires
                self.module_buildrequires == config.get('buildrequires', {})
            )):
                found_config = config
                break

        if found_config is None:
            log.warning("Specified module is not found under tag %s in tag config file, "
                        "nothing to change", self.tag)
        else:
            module_configs.remove(found_config)
            log.info("Updating tag config file to remove module:\n%r", found_config)
            if not self.dry_run:
                self.write_tag_config()
